requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework_simplejwt
mysqlclient
requests
pandas
Pillow